UI test cleanup: Extract derive_hash_xor_eq tests
